refactor: type과의 싸움에서 지다
