+++ client/src/types/type.ts
@@ -41,3 +41,7 @@ export interface IButtonStyle {
   onClick?: () => void;
   onSubmit?: () => void;
 }
+
+export type MarkerT = {
+  markerId: number | null;
+};
